// src/front/src/api/rcp/models/monthlyEntrySummaryResponse.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * PBX Rcp API (via Gateway)
 * OpenAPI spec version: 1.0
 */
import type { RcpTimeEntryStatus } from "./rcpTimeEntryStatus";

export interface MonthlyEntrySummaryResponse {
	/** @nullable */
	gid?: string | null;
	/** @nullable */
	employeeFullName?: string | null;
	/** @nullable */
	userGid?: string | null;
	status?: RcpTimeEntryStatus;
	totalMinutes?: number;
	/** @nullable */
	totalFormatted?: string | null;
	/** @nullable */
	submittedAt?: string | null;
}